# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlimefishBackend API",
    "version": "1.0",
    "description": "[Discord](https://discord.gg/Q5CeSMFeBP) · [API key](https://slimefishbackend.dev/settings/api) · [OpenAPI spec](https://github.com/casesandberg/slimefish-backend/blob/main/apps/api/openapi.json) · [Trading bot](https://github.com/casesandberg/slimefishbackend-bot-trading) · [Mirror bot](https://github.com/casesandberg/slimefishbackend-bot-manifold-mirror)\n\n Include `x-api-key: {{key}}` in the header of your requests."
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "cursor": {
        "type": "string"
      },
      "limit": {
        "type": "number",
        "minimum": 1,
        "maximum": 100
      },
      "sortField": {
        "type": "string"
      },
      "sortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
      "status": {
        "type": "string",
        "enum": [
          "active",
          "closed",
          "all"
        ]
      },
      "createdBy": {
        "type": "string"
      },
      "tags": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
...
    },
    "parameters": {
...
      "cursor": {
        "schema": {
          "$ref": "#/components/schemas/cursor"
        },
        "required": false,
        "name": "cursor",
        "in": "path"
      },
      "limit": {
        "schema": {
          "$ref": "#/components/schemas/limit"
        },
        "required": false,
        "name": "limit",
        "in": "path"
      },
      "sortField": {
        "schema": {
          "$ref": "#/components/schemas/sortField"
        },
        "required": false,
        "name": "sortField",
        "in": "path"
      },
      "sortDirection": {
        "schema": {
          "$ref": "#/components/schemas/sortDirection"
        },
        "required": false,
        "name": "sortDirection",
        "in": "path"
      },
...
      "status": {
        "schema": {
          "$ref": "#/components/schemas/status"
        },
        "required": false,
        "name": "status",
        "in": "path"
      },
      "createdBy": {
        "schema": {
          "$ref": "#/components/schemas/createdBy"
        },
        "required": false,
        "name": "createdBy",
        "in": "path"
      },
      "tags": {
        "schema": {
          "$ref": "#/components/schemas/tags"
        },
        "required": false,
        "name": "tags",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/markets": {
      "get": {
        "summary": "Get markets",
        "parameters": [
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/createdBy"
          },
          {
            "$ref": "#/components/parameters/tags"
          },
          {
            "$ref": "#/components/parameters/cursor"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/sortField"
          },
          {
            "$ref": "#/components/parameters/sortDirection"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "cuid"
                          },
                          "question": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string",
                            "minLength": 1
                          },
                          "closeDate": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "resolvedAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "canceledAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "canceledById": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "createdBy": {
                            "type": "string"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            },
                            "maxItems": 5
                          },
                          "ammAccountId": {
                            "type": "string"
                          },
                          "clearingAccountId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "updatedAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "commentCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "uniqueTradersCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "uniquePromotersCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "liquidityCount": {
                            "type": [
                              "integer",
                              "null"
                            ]
                          },
                          "parentListId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "question",
                          "description",
                          "slug",
                          "closeDate",
                          "resolvedAt",
                          "canceledAt",
                          "canceledById",
                          "createdBy",
                          "tags",
                          "ammAccountId",
                          "clearingAccountId",
                          "createdAt",
                          "updatedAt",
                          "commentCount",
                          "uniqueTradersCount",
                          "uniquePromotersCount",
                          "liquidityCount",
                          "parentListId"
                        ]
                      }
                    },
                    "pageInfo": {
                      "type": "object",
                      "properties": {
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "endCursor": {
                          "type": "string"
                        },
                        "total": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "hasNextPage",
                        "total"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "pageInfo"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a market",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "closeDate": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 5
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "color": {
                          "type": "string",
                          "pattern": "^#[0-9A-Fa-f]{6}$"
                        }
                      },
                      "required": [
                        "name",
                        "color"
                      ]
                    }
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "binary",
                      "multi",
                      "list"
                    ]
                  },
                  "contributionPolicy": {
                    "type": "string",
                    "enum": [
                      "DISABLED",
                      "OWNERS_ONLY",
                      "FRIENDS_ONLY",
                      "PUBLIC"
                    ]
                  }
                },
                "required": [
                  "question",
                  "description",
                  "closeDate",
                  "tags",
                  "options",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "market": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "cuid"
                            },
                            "question": {
                              "type": "string",
                              "minLength": 1
                            },
                            "description": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string",
                              "minLength": 1
                            },
                            "closeDate": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "resolvedAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "canceledAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "canceledById": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "createdBy": {
                              "type": "string"
                            },
                            "tags": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              },
                              "maxItems": 5
                            },
                            "ammAccountId": {
                              "type": "string"
                            },
                            "clearingAccountId": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "updatedAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "commentCount": {
                              "type": [
                                "integer",
                                "null"
                              ]
                            },
                            "uniqueTradersCount": {
                              "type": [
                                "integer",
                                "null"
                              ]
                            },
                            "uniquePromotersCount": {
                              "type": [
                                "integer",
                                "null"
                              ]
                            },
                            "liquidityCount": {
                              "type": [
                                "integer",
                                "null"
                              ]
                            },
                            "parentListId": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "question",
                            "description",
                            "slug",
                            "closeDate",
                            "resolvedAt",
                            "canceledAt",
                            "canceledById",
                            "createdBy",
                            "tags",
                            "ammAccountId",
                            "clearingAccountId",
                            "createdAt",
                            "updatedAt",
                            "commentCount",
                            "uniqueTradersCount",
                            "uniquePromotersCount",
                            "liquidityCount",
                            "parentListId"
                          ]
                        },
                        "list": {
                          "type": "object",
                          "properties": {
                            "contributionPolicy": {
                              "type": "string",
                              "enum": [
                                "DISABLED",
                                "OWNERS_ONLY",
                                "FRIENDS_ONLY",
                                "PUBLIC"
                              ]
                            },
                            "id": {
                              "type": "string",
                              "format": "cuid"
                            },
                            "title": {
                              "type": "string",
                              "minLength": 1
                            },
                            "slug": {
                              "type": "string",
                              "minLength": 1
                            },
                            "description": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "ownerId": {
                              "type": "string"
                            },
                            "contributionReview": {
                              "type": [
                                "boolean",
                                "null"
                              ]
                            },
                            "tags": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              },
                              "maxItems": 5
                            },
                            "createdAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "updatedAt": {
                              "type": [
                                "string",
                                "null"
                              ]
                            }
                          },
                          "required": [
                            "contributionPolicy",
                            "id",
                            "title",
                            "slug",
                            "description",
                            "ownerId",
                            "contributionReview",
                            "tags",
                            "createdAt",
                            "updatedAt"
                          ]
                        }
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}